+++ pytorch_widedeep/trainer/__init__.py
@@ -1,0 +1,1 @@
+from .trainer import Trainer  # noqa: F401
